inte/networking: move port mapping tests into a dedicated file
